[google_sign_in] Convert Android to Pigeon (#4344)
Replaces the direct method channel implementation with Pigeon.

Since `google_sign_in`, unlike most of our plugins, exposes an API that's intended for direct cross-plugin native use, the existing methods are all left in place, but refactored as passthroughs to the new Pigeon versions. To ensure that they aren't broken, the existing Java tests are preserved unchanged (as a "legacy" copy) with `onMethodCall` left in place for now just to allow the tests to continue to run as-is. Since that dispatches to the legacy methods, this keeps the existing coverage of those methods.

The new tests are a copy of the legacy tests, minimally translated to use the new Pigeon variants, to ensure continuity of testing to the new version.

Part of https://github.com/flutter/flutter/issues/117908