[camera_android]: Fixes crash when record video on Android versions lower than 12 (#4635)
The the solution of the issue flutter/flutter#109769 in the #7073 causes a crashes when the Android SDK version is lower than 31, due to EncoderProfiles not being found on these devices. This issue was solved putting the EncoderProfiles declaration after the check for SDK version.

based on https://github.com/flutter/plugins/pull/7073 by @camsim99.

Fixes https://github.com/flutter/flutter/issues/109769 in Android lower than 12.