[ios_platform_images] Add integration tests (#4899)
Currently there are no end-to-end tests for this package that would fail if the communication with the native implementation didn't work, or if the native implementation were broken. This is especially problematic since we want to convert this plugin to both Swift and Pigeon, each of which has the potential to break things that would currently not be caught.

This adds basic integration tests of both methods to ensure that the simple case of a successful call works as expected.