[flutter_markdown] Fix changelog regarding minimum supported SDK version (#4851)
The changelog of the `flutter_markdown` package is a bit confusing. First it says:

> Updates minimum supported SDK version to Flutter 3.10/Dart 3.0

and in the latest update is says

> Updates minimum supported SDK version to Flutter 3.7/Dart 2.19

Just from reading this I would expect that the minimum SDK lowered with the latest update. However, this is not the case.

After a bit of researching, I found this was caused by https://github.com/flutter/packages/pull/4731. The mentioned PR only updates the minimum SDK of the example but not of the package, see: https://github.com/flutter/packages/pull/4731/files#diff-e780ee3aec5fdc10512b452b5464878a11caf22fb23d28f725324308ad243d17R6

Therefore, I clarified the changelog with this PR.

![telegram-cloud-photo-size-2-5456132896298487892-y](https://github.com/flutter/flutter/assets/24459435/c992917a-7eda-4e45-a656-50408a6c8910)

<img width="819" alt="image" src="https://github.com/flutter/flutter/assets/24459435/bee2f046-ea86-45c7-9567-a19949e3e6f8">

Fixes https://github.com/flutter/flutter/issues/134056

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*