[url_launcher] Adds documentation that a launch needs to be triggered  by a user action (#5143)
Some web browsers, such as Safari, may prevent URL launching if it is not triggered by a user action (e.g. a button click). Even if a user triggers a launch through a button click, if there is a delay due to awaiting a Future before the launch, the browser may still block it. This can be confusing if you don't know it. As https://github.com/flutter/flutter/issues/78524 suggested, should this be documented. Therefore, I added a few sentences to the `launchUrl`.

Please let me know if this documentation is too long or not the right place.

Fixes https://github.com/flutter/flutter/issues/78524