[go_router] Expose full `Uri` on `GoRouterState` in `GoRouterRedirect` (#5742)
A number of developers have voiced the desire to be able to know whether a `GoRouterState` is the result of a deep-link or in-app navigation from within their `GoRouter`'s `redirect` method. This can be accomplished by exposing the `Uri`'s `scheme` and `host` on the `GoRouterState`. This way, we can know whether the `GoRouterState` is the result of a deep-link by checking `state.uri.scheme != null` or `state.uri.host != null`.

This PR would close [#103659](https://github.com/flutter/flutter/issues/103659#issuecomment-1865112881).

No tests were broken as a result of this change, and we have added coverage for this change through new tests.