[two_dimensional_scrollables] Refactor Spans for common use (#6550)
This relocates most of `table_span.dart` to `../common/span.dart`, renaming the classes to basic Spans instead of TableSpans.
This is because since a Span can represent a row or a column, the new base Span class will be used to build TreeRows in the upcoming TreeView.
In `table_span.dart`, the old classes are now typedefs. This is because (IMO) the Table and Tree APIs are better to work with when the naming conventions reflect what you are building. The Table and Tree uses of Span could also deviate in the future and it felt prudent to leave room for future changes that may be specific to only one or the other.

Otherwise a bit of tidying to get ready for TreeView!

Part of https://github.com/flutter/flutter/issues/42332