[google_maps_flutter] Add iOS SDK 9.x support (#6902)
Expands the range of supported versions of the iOS Google Maps SDK to include 9.x. To ensure build compatibility coverage, this follows the previously-established model of making duplicate example apps with different min iOS versions, but without most of the tests (as described in example/README.md).

The new `ios15` example directory is a duplicate of `ios14/` with:
- The min project and Podfile versions upped to 15.0 (I've validated locally that this causes a build to get 9.0 of the SDK).
- The integration tests and XCUITest are removed.
- All but the main plugin XCTest tests are removed (that file is left mainly as a placeholder to add new unit tests if necessary).

This does not remove the privacy manifest workaround for 8.4, as we still support 8.4; the resulting duplication should be harmless for now. See discussion in https://github.com/flutter/flutter/issues/146838.

Fixes https://github.com/flutter/flutter/issues/149873